backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "bs4>=0.0.2",
    "cohere>=5.20.0",
    "fastapi[standard]>=0.124.4",
    "langchain-text-splitters>=1.1.0",
    "lxml>=6.0.2",
    "openai-agents>=0.6.3",
    "python-dotenv>=1.2.1",
    "qdrant-client>=1.16.2",
    "requests",
]
